refactor(ui): convert EventStream from drawer to panel
Convert the EventStreamDrawer to EventStreamPanel to allow non-modal
interaction with the app while viewing live events.

Changes:
- Renamed EventStreamDrawer to EventStreamPanel
- Changed from modal Drawer to collapsible side panel (similar to CommentsPanel)
- Panel appears on the right side of the app layout
- Persists collapsed state to localStorage (agor:eventStreamPanelCollapsed)
- Only captures events when panel is expanded (performance optimization)
- Users can now interact with the app while monitoring events

The panel can be toggled via the navbar icon and remembers its state
across sessions.
