--- apps/agor-ui/src/components/EventStreamDrawer/index.ts
@@ -1,4 +1,0 @@
-export type { EventItemProps } from './EventItem';
-export { EventItem } from './EventItem';
-export type { EventStreamDrawerProps } from './EventStreamDrawer';
-export { EventStreamDrawer } from './EventStreamDrawer';
