+++ apps/agor-ui/src/components/EventStreamPanel/EventItem.tsx
